pyproject.toml
[tool.poetry]
name = "ai-automatic-env-build"
version = "0.1.0"
description = ""
authors = ["stevechan <chanshunli@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11, <3.12"
poetry = "^1.8.3"
llama-index-core = "^0.10.57"
llama-index-llms-openai = "^0.1.26"
llama-index-llms-replicate = "^0.1.3"
llama-index-embeddings-huggingface = "^0.2.2"
python-dotenv = "^1.0.1"
llama-index-llms-ollama = "^0.1.6"
llama-index-readers-file = "^0.1.30"
llama-index-embeddings-openai = "^0.1.11"
llama-parse = "^0.4.9"
langfuse = "^2.36.1"
socksio = "^1.0.0"
ipython = "^8.29.0"
aiohttp = "^3.11.7"
pyautogen = "^0.3.2"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

